+++ Classes/Exception/NoSuchHaikuException.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace T3docs\Examples\Exception;
+
+/**
+ * Exception thrown if a haiku record was not found
+ */
+class NoSuchHaikuException extends \RuntimeException
+{
+}