+++ Configuration/TCA/Overrides/tt_content_plugin_haiku_detail.php
@@ -1,0 +1,22 @@
+<?php
+
+defined('TYPO3') or die();
+
+\TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addPlugin(
+    [
+        'LLL:EXT:examples/Resources/Private/Language/PluginHaiku/locallang_db.xlf:detail.title',
+        'examples_haiku_detail',
+        'tx_examples-haiku',
+    ],
+    'list_type',
+    'examples'
+);
+
+$GLOBALS['TCA']['tt_content']['types']['list']['subtypes_excludelist']['examples_haiku_detail']='pages,layout,select_key,recursive';
+
+$GLOBALS['TCA']['tt_content']['types']['list']['subtypes_addlist']['examples_haiku_detail']='pi_flexform';
+
+\TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addPiFlexFormValue(
+    'examples_haiku_detail',
+    'FILE:EXT:examples/Configuration/Flexforms/PluginHaikuDetail.xml'
+);